move gui style related configurable options to dedicated unit and update icons when inversion changed
